requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu
anyio==3.5.0
Flask==2.1.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pip==22.0.4
pyTelegramBotAPI==4.4.0
telebot==0.0.4
